Refactor into a single source file.
This is a pretty major refactor of the code so that mangle is easier to
pull into an existing project. All of the source code for Mangle itself
has been pulled into mangle.c, and there is now just a single batch file
and a shell script (the latter untested) which can be deployed with that
source file to run things.