Add a basic Makefile.
During development, it is nice to have a Makefile that I can use for iterating on the code. This makefile captures the idiom of generating a "bootstrap" version of the tool from the checked in `.c` file, and then iterating by using that tool to generate the `.c` from all the `.md` files.

Anybody who uses Mangle shouldn't need this step, since they can just use `mangle.sh` or `mangle.bat`, which will build things on demand anyway.